+++ lib/production/validations/work_order_item_schema.rb
@@ -5,8 +5,8 @@ module ProductionApp
     optional(:id).filled(:integer)
     optional(:work_order_id).filled(:integer)
     optional(:product_setup_id).filled(:integer)
-    required(:carton_qty_required).maybe(:decimal)
-    optional(:carton_qty_produced).maybe(:decimal)
+    required(:carton_qty_required).maybe(:integer)
+    optional(:carton_qty_produced).maybe(:integer)
     optional(:completed).maybe(:bool)
     optional(:completed_at).maybe(:time)
   end
